refactor: helpers para github movido y hooks para botones movidos
